Move SearchActivity to correct package
